refactor: rename test file to SchemaManagerIT as now a class.
